Updated py_srvcli and sensors for service/client that gets depth

+++ competitionCode2024/ros2_ws/install/sensors/include/sensors/sensors/msg/detail/control_data__struct.h
@@ -25,7 +25,7 @@ extern "C"
 typedef struct sensors__msg__ControlData
 {
   sensor_msgs__msg__Imu imu_data;
-  int64_t depth;
+  double depth;
 } sensors__msg__ControlData;
 
 // Struct for a sequence of sensors__msg__ControlData.
